Split apart TestProvider from react-testing and remove react-testing from bundle
